+++ src/ext/java/org/opentripplanner/ext/vectortiles/layers/vehiclerental/DigitransitVehicleRentalPropertyMapper.java
@@ -16,9 +16,9 @@ public class DigitransitVehicleRentalPropertyMapper extends PropertyMapper<Vehic
   @Override
   protected Collection<T2<String, Object>> map(VehicleRentalStation station) {
     return List.of(
-        new T2<>("id", station.id),
+        new T2<>("id", station.getStationId()),
         new T2<>("name", station.name),
-        new T2<>("networks", String.join(",", station.networks))
+        new T2<>("networks", station.getNetwork())
     );
   }
 }