+++ src/main/java/org/opentripplanner/api/mapping/VehicleRentalStationMapper.java
@@ -3,7 +3,7 @@ package org.opentripplanner.api.mapping;
 import org.opentripplanner.api.model.ApiVehicleRentalStation;
 import org.opentripplanner.routing.vehicle_rental.VehicleRentalStation;
 
-import java.util.ArrayList;
+import java.util.List;
 import java.util.Locale;
 
 public class VehicleRentalStationMapper {
@@ -12,7 +12,7 @@ public class VehicleRentalStationMapper {
 
         ApiVehicleRentalStation api = new ApiVehicleRentalStation();
 
-        api.id = domain.id;
+        api.id = domain.getStationId();
         api.name = domain.name.toString(locale);
         api.x = domain.longitude;
         api.y = domain.latitude;
@@ -21,7 +21,7 @@ public class VehicleRentalStationMapper {
         api.allowDropoff = domain.allowDropoff;
         api.isFloatingBike = domain.isFloatingBike;
         api.isCarStation = domain.isCarStation;
-        api.networks = new ArrayList<>(domain.networks);
+        api.networks = List.of(domain.getNetwork());
         api.realTimeData = domain.realTimeData;
         api.rentalUris = domain.rentalUris;