+++ src/main/java/org/opentripplanner/routing/vehicle_rental/VehicleRentalStationService.java
@@ -11,17 +11,22 @@ import java.util.stream.Collectors;
 
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Envelope;
+import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.routing.bike_park.BikePark;
 
 public class VehicleRentalStationService implements Serializable {
     private static final long serialVersionUID = -1288992939159246764L;
 
-    private final Map<String, VehicleRentalStation> vehicleRentalStations = new HashMap<>();
+    private final Map<FeedScopedId, VehicleRentalStation> vehicleRentalStations = new HashMap<>();
 
     private Set<BikePark> bikeParks = new HashSet<>();
 
     public Collection<VehicleRentalStation> getVehicleRentalStations() {
         return vehicleRentalStations.values();
+    }
+
+    public VehicleRentalStation getVehicleRentalStation(FeedScopedId id) {
+        return vehicleRentalStations.get(id);
     }
 
     public void addVehicleRentalStation(VehicleRentalStation vehicleRentalStation) {
@@ -30,7 +35,7 @@ public class VehicleRentalStationService implements Serializable {
         vehicleRentalStations.put(vehicleRentalStation.id, vehicleRentalStation);
     }
 
-    public void removeVehicleRentalStation(String vehicleRentalStationId) {
+    public void removeVehicleRentalStation(FeedScopedId vehicleRentalStationId) {
         vehicleRentalStations.remove(vehicleRentalStationId);
     }